+++ dictionaries/el/cspell-tools.config.yaml
@@ -1,0 +1,10 @@
+# yaml-language-server: $schema=https://raw.githubusercontent.com/streetsidesoftware/cspell/main/packages/cspell-tools/cspell-tools.config.schema.json
+
+targets:
+  - name: el_GR
+    targetDirectory: .
+    compress: false
+    format: trie3
+    sources:
+      - listFile: source-files.txt
+    generateNonStrict: true
