+++ dictionaries/el/cspell.json
@@ -9,7 +9,7 @@
     "ignorePaths": [
         "/src"
     ],
-    "ignoreWords": [],
+    "ignoreWords": ["ALLCAPS"],
     "overrides": [
         {
             "filename": "/src",
