+++ dictionaries/el/package.json
@@ -11,10 +11,12 @@
     "./cspell-ext.json": "./cspell-ext.json"
   },
   "scripts": {
-    "compile-all-caps": "node ./tools/toCaps ./src/Greek.txt ./src/Greek.ALLCAPS.txt",
-    "build-dictionary": "cross-env NODE_OPTIONS=--max_old_space_size=4096 cspell-tools compile-trie --merge el_GR src/hunspell/el-GR.dic src/Greek.txt src/Greek.ALLCAPS.txt -o .",
-    "build": "pnpm run compile-all-caps && pnpm run build-dictionary",
-    "prepublishOnly": "echo pre-publish",
+    "compile-all-caps": "node ./tools/toCaps.mjs ./src/Greek.txt ./src/Greek.ALLCAPS.txt",
+    "build:dictionary:old": "cross-env NODE_OPTIONS=--max_old_space_size=4096 cspell-tools-cli compile --trie3 -x compound --merge el_GR src/hunspell/el-GR.dic src/Greek.txt src/Greek.ALLCAPS.txt -o .",
+    "build:dictionary": "cspell-tools-cli build",
+    "build:gz": "cspell-tools-cli gzip *.trie",
+    "build": "pnpm run compile-all-caps && pnpm run build:dictionary && pnpm build:gz",
+    "prepare": "pnpm run build:gz",
     "test": "shx head -n 1000 ./src/Greek.txt | cspell -v -c cspell-ext.json --local=el stdin"
   },
   "repository": {
