+++ dictionaries/el/source-files.txt
@@ -1,0 +1,3 @@
+src/hunspell/el-GR.dic
+src/Greek.txt
+src/Greek.ALLCAPS.txt
