+++ package.json
@@ -29,7 +29,7 @@
     "url": "https://github.com/streetsidesoftware/cspell-dicts.git"
   },
   "devDependencies": {
-    "@cspell/cspell-tools": "^7.0.1-alpha.6",
+    "@cspell/cspell-tools": "^7.0.1-alpha.7",
     "@lerna-lite/cli": "^2.5.0",
     "@lerna-lite/exec": "^2.5.0",
     "@lerna-lite/publish": "^2.5.0",
